+++ .pypirc
@@ -1,0 +1,14 @@
+[distutils]
+index-servers =
+  pypi
+  testpypi
+
+[pypi]
+repository: https://upload.pypi.org/legacy/
+username: ikrfun
+password: Ikrfun130625!
+
+[testpypi]
+repository: https://test.pypi.org/legacy/
+username: ikrfun
+password: Ikrfun130625!
