+++ requirements.txt
@@ -1,4 +1,19 @@
-numpy 
-pandas
-opencv-python
-tqdm
+matplotlib>=3.7.1
+matplotlib>=3.7.0
+moviepy>=1.0.3
+numpy>=1.24.3
+numpy>=1.23.5
+opencv_contrib_python>=4.6.0.66
+opencv_python>=4.5.5.64
+opencv_python>=4.7.0.72
+opencv_python_headless>=4.7.0.72
+pandas>=2.0.1
+pandas>=2.0.2
+pandas>=1.5.3
+setuptools>=67.7.2
+setuptools>=66.0.0
+setuptools>=67.8.0
+setuptools>=65.6.3
+tqdm>=4.65.0
+tqdm>=4.64.1
+utils>=1.0.1
