--- valorant_urils/coords/coord_data.json
@@ -1,1 +1,0 @@
-{"minimap": [[4, 0], [3, 4], [4, 3], [3, 1], [3, 7], [5, 4], [4, 6], [5, 1], [5, 7], [0, 2], [0, 5], [2, 2], [1, 6], [2, 5], [1, 3], [7, 4], [6, 2], [6, 5], [4, 2], [3, 0], [4, 5], [3, 3], [5, 6], [3, 6], [5, 3], [2, 4], [1, 2], [0, 4], [2, 1], [2, 7], [1, 5], [6, 1], [6, 4], [7, 3], [3, 2], [4, 1], [4, 7], [3, 5], [5, 2], [4, 4], [5, 5], [1, 1], [0, 3], [2, 0], [1, 4], [0, 6], [2, 3], [1, 7], [2, 6], [6, 6], [7, 5], [6, 3]], "left_team": [[11, 1], [9, 0], [12, 1], [7, 1], [13, 1], [8, 1], [11, 0], [12, 0], [7, 0], [10, 1], [13, 0], [8, 0], [10, 0], [9, 1]], "time": [[15, 0], [15, 1], [16, 0], [16, 1]], "right_team": [[23, 1], [18, 1], [21, 0], [24, 1], [19, 1], [22, 1], [20, 1], [18, 0], [23, 0], [24, 0], [19, 0], [20, 0], [22, 0], [21, 1]], "hp": [[9, 16], [9, 17], [10, 17], [10, 16]], "skill": [[17, 17], [12, 17], [18, 17], [13, 17], [16, 16], [14, 17], [17, 16], [12, 16], [15, 17], [18, 16], [13, 16], [19, 17], [14, 16], [19, 16], [15, 16], [16, 17]], "ammo": [[21, 16], [21, 17], [22, 17], [22, 16]], "centor": [[16, 8], [16, 9], [15, 8], [15, 9]], "killlog": [[27, 1], [27, 4], [28, 3], [30, 3], [26, 2], [29, 4], [29, 1], [28, 2], [27, 3], [30, 2], [26, 1], [26, 4], [29, 3], [28, 4], [27, 2], [28, 1], [30, 4], [30, 1], [26, 3], [29, 2]]}
